+++ datasets/cpp-keep-code_davinci_001_temp_0.2/HumanEval_116_sort_array.cpp
@@ -52,12 +52,12 @@ std::vector<long> sort_array(std::vector<long> arr) {
 }
 int main() {
     auto candidate = sort_array;
-    assert(candidate((std::vector<long>({1, 5, 2, 3, 4}))) == (std::vector<long>({1, 2, 4, 3, 5})));
-    assert(candidate((std::vector<long>({-2, -3, -4, -5, -6}))) == (std::vector<long>({-4, -2, -6, -5, -3})));
-    assert(candidate((std::vector<long>({1, 0, 2, 3, 4}))) == (std::vector<long>({0, 1, 2, 4, 3})));
+    assert(candidate((std::vector<long>({(long)1, (long)5, (long)2, (long)3, (long)4}))) == (std::vector<long>({(long)1, (long)2, (long)4, (long)3, (long)5})));
+    assert(candidate((std::vector<long>({(long)-2, (long)-3, (long)-4, (long)-5, (long)-6}))) == (std::vector<long>({(long)-4, (long)-2, (long)-6, (long)-5, (long)-3})));
+    assert(candidate((std::vector<long>({(long)1, (long)0, (long)2, (long)3, (long)4}))) == (std::vector<long>({(long)0, (long)1, (long)2, (long)4, (long)3})));
     assert(candidate((std::vector<long>())) == (std::vector<long>()));
-    assert(candidate((std::vector<long>({2, 5, 77, 4, 5, 3, 5, 7, 2, 3, 4}))) == (std::vector<long>({2, 2, 4, 4, 3, 3, 5, 5, 5, 7, 77})));
-    assert(candidate((std::vector<long>({3, 6, 44, 12, 32, 5}))) == (std::vector<long>({32, 3, 5, 6, 12, 44})));
-    assert(candidate((std::vector<long>({2, 4, 8, 16, 32}))) == (std::vector<long>({2, 4, 8, 16, 32})));
-    assert(candidate((std::vector<long>({2, 4, 8, 16, 32}))) == (std::vector<long>({2, 4, 8, 16, 32})));
+    assert(candidate((std::vector<long>({(long)2, (long)5, (long)77, (long)4, (long)5, (long)3, (long)5, (long)7, (long)2, (long)3, (long)4}))) == (std::vector<long>({(long)2, (long)2, (long)4, (long)4, (long)3, (long)3, (long)5, (long)5, (long)5, (long)7, (long)77})));
+    assert(candidate((std::vector<long>({(long)3, (long)6, (long)44, (long)12, (long)32, (long)5}))) == (std::vector<long>({(long)32, (long)3, (long)5, (long)6, (long)12, (long)44})));
+    assert(candidate((std::vector<long>({(long)2, (long)4, (long)8, (long)16, (long)32}))) == (std::vector<long>({(long)2, (long)4, (long)8, (long)16, (long)32})));
+    assert(candidate((std::vector<long>({(long)2, (long)4, (long)8, (long)16, (long)32}))) == (std::vector<long>({(long)2, (long)4, (long)8, (long)16, (long)32})));
 }
