+++ datasets/cpp-keep-code_davinci_001_temp_0.2/HumanEval_121_solution.cpp
@@ -17,11 +17,11 @@ long solution(std::vector<long> lst) {
 }
 int main() {
     auto candidate = solution;
-    assert(candidate((std::vector<long>({5, 8, 7, 1}))) == (12));
-    assert(candidate((std::vector<long>({3, 3, 3, 3, 3}))) == (9));
-    assert(candidate((std::vector<long>({30, 13, 24, 321}))) == (0));
-    assert(candidate((std::vector<long>({5, 9}))) == (5));
-    assert(candidate((std::vector<long>({2, 4, 8}))) == (0));
-    assert(candidate((std::vector<long>({30, 13, 23, 32}))) == (23));
-    assert(candidate((std::vector<long>({3, 13, 2, 9}))) == (3));
+    assert(candidate((std::vector<long>({(long)5, (long)8, (long)7, (long)1}))) == (12));
+    assert(candidate((std::vector<long>({(long)3, (long)3, (long)3, (long)3, (long)3}))) == (9));
+    assert(candidate((std::vector<long>({(long)30, (long)13, (long)24, (long)321}))) == (0));
+    assert(candidate((std::vector<long>({(long)5, (long)9}))) == (5));
+    assert(candidate((std::vector<long>({(long)2, (long)4, (long)8}))) == (0));
+    assert(candidate((std::vector<long>({(long)30, (long)13, (long)23, (long)32}))) == (23));
+    assert(candidate((std::vector<long>({(long)3, (long)13, (long)2, (long)9}))) == (3));
 }
