+++ datasets/cpp-keep-code_davinci_001_temp_0.2/HumanEval_136_largest_smallest_integers.cpp
@@ -26,15 +26,15 @@ std::tuple<std::optional<long>, std::optional<long>> largest_smallest_integers(s
 }
 int main() {
     auto candidate = largest_smallest_integers;
-    assert(candidate((std::vector<long>({2, 4, 1, 3, 5, 7}))) == std::make_tuple(std::optional<long>({}), std::optional<long>(1)));
-    assert(candidate((std::vector<long>({2, 4, 1, 3, 5, 7, 0}))) == std::make_tuple(std::optional<long>({}), std::optional<long>(1)));
-    assert(candidate((std::vector<long>({1, 3, 2, 4, 5, 6, -2}))) == std::make_tuple(-2, 1));
-    assert(candidate((std::vector<long>({4, 5, 3, 6, 2, 7, -7}))) == std::make_tuple(-7, 2));
-    assert(candidate((std::vector<long>({7, 3, 8, 4, 9, 2, 5, -9}))) == std::make_tuple(-9, 2));
+    assert(candidate((std::vector<long>({(long)2, (long)4, (long)1, (long)3, (long)5, (long)7}))) == std::make_tuple(std::optional<long>({}), std::optional<long>(1)));
+    assert(candidate((std::vector<long>({(long)2, (long)4, (long)1, (long)3, (long)5, (long)7, (long)0}))) == std::make_tuple(std::optional<long>({}), std::optional<long>(1)));
+    assert(candidate((std::vector<long>({(long)1, (long)3, (long)2, (long)4, (long)5, (long)6, (long)-2}))) == std::make_tuple(-2, 1));
+    assert(candidate((std::vector<long>({(long)4, (long)5, (long)3, (long)6, (long)2, (long)7, (long)-7}))) == std::make_tuple(-7, 2));
+    assert(candidate((std::vector<long>({(long)7, (long)3, (long)8, (long)4, (long)9, (long)2, (long)5, (long)-9}))) == std::make_tuple(-9, 2));
     assert(candidate((std::vector<long>())) == std::make_tuple(std::optional<long>({}), std::optional<long>({})));
-    assert(candidate((std::vector<long>({0}))) == std::make_tuple(std::optional<long>({}), std::optional<long>({})));
-    assert(candidate((std::vector<long>({-1, -3, -5, -6}))) == std::make_tuple(std::optional<long>(-1), std::optional<long>({})));
-    assert(candidate((std::vector<long>({-1, -3, -5, -6, 0}))) == std::make_tuple(std::optional<long>(-1), std::optional<long>({})));
-    assert(candidate((std::vector<long>({-6, -4, -4, -3, 1}))) == std::make_tuple(-3, 1));
-    assert(candidate((std::vector<long>({-6, -4, -4, -3, -100, 1}))) == std::make_tuple(-3, 1));
+    assert(candidate((std::vector<long>({(long)0}))) == std::make_tuple(std::optional<long>({}), std::optional<long>({})));
+    assert(candidate((std::vector<long>({(long)-1, (long)-3, (long)-5, (long)-6}))) == std::make_tuple(std::optional<long>(-1), std::optional<long>({})));
+    assert(candidate((std::vector<long>({(long)-1, (long)-3, (long)-5, (long)-6, (long)0}))) == std::make_tuple(std::optional<long>(-1), std::optional<long>({})));
+    assert(candidate((std::vector<long>({(long)-6, (long)-4, (long)-4, (long)-3, (long)1}))) == std::make_tuple(-3, 1));
+    assert(candidate((std::vector<long>({(long)-6, (long)-4, (long)-4, (long)-3, (long)-100, (long)1}))) == std::make_tuple(-3, 1));
 }
