+++ datasets/cpp-keep-code_davinci_001_temp_0.2/HumanEval_52_below_threshold.cpp
@@ -16,10 +16,10 @@ bool below_threshold(std::vector<long> l, long t) {
 }
 int main() {
     auto candidate = below_threshold;
-    assert(candidate((std::vector<long>({1, 2, 4, 10})), (100)) == (true));
-    assert(candidate((std::vector<long>({1, 20, 4, 10})), (5)) == (false));
-    assert(candidate((std::vector<long>({1, 20, 4, 10})), (21)) == (true));
-    assert(candidate((std::vector<long>({1, 20, 4, 10})), (22)) == (true));
-    assert(candidate((std::vector<long>({1, 8, 4, 10})), (11)) == (true));
-    assert(candidate((std::vector<long>({1, 8, 4, 10})), (10)) == (false));
+    assert(candidate((std::vector<long>({(long)1, (long)2, (long)4, (long)10})), (100)) == (true));
+    assert(candidate((std::vector<long>({(long)1, (long)20, (long)4, (long)10})), (5)) == (false));
+    assert(candidate((std::vector<long>({(long)1, (long)20, (long)4, (long)10})), (21)) == (true));
+    assert(candidate((std::vector<long>({(long)1, (long)20, (long)4, (long)10})), (22)) == (true));
+    assert(candidate((std::vector<long>({(long)1, (long)8, (long)4, (long)10})), (11)) == (true));
+    assert(candidate((std::vector<long>({(long)1, (long)8, (long)4, (long)10})), (10)) == (false));
 }
